librz/lang: split RzLang in plugins
