+++ librz/lang/lang.c
@@ -3,19 +3,13 @@
 
 #include <rz_lang.h>
 #include <rz_util.h>
+#include <config.h>
 
 RZ_LIB_VERSION(rz_lang);
 
-#include "p/pipe.c" // hardcoded
-#include "p/vala.c" // hardcoded
-#include "p/rust.c" // hardcoded
-#include "p/zig.c" // hardcoded
-#include "p/spp.c" // hardcoded
-#include "p/c.c" // hardcoded
-#include "p/lib.c"
-#if __UNIX__
-#include "p/cpipe.c" // hardcoded
-#endif
+static RzLangPlugin *lang_static_plugins[] = {
+	RZ_LANG_STATIC_PLUGINS
+};
 
 static RzLang *__lang = NULL;
 
@@ -44,16 +38,9 @@ RZ_API RzLang *rz_lang_new(void) {
 	}
 	lang->defs->free = (RzListFree)rz_lang_def_free;
 	lang->cb_printf = (PrintfCallback)printf;
-#if __UNIX__
-	rz_lang_add(lang, &rz_lang_plugin_c);
-	rz_lang_add(lang, &rz_lang_plugin_cpipe);
-#endif
-	rz_lang_add(lang, &rz_lang_plugin_vala);
-	rz_lang_add(lang, &rz_lang_plugin_rust);
-	rz_lang_add(lang, &rz_lang_plugin_zig);
-	rz_lang_add(lang, &rz_lang_plugin_spp);
-	rz_lang_add(lang, &rz_lang_plugin_pipe);
-	rz_lang_add(lang, &rz_lang_plugin_lib);
+	for (int i = 0; lang_static_plugins[i]; i++) {
+		rz_lang_add(lang, lang_static_plugins[i]);
+	}
 
 	return lang;
 }
